[3/3] Rename service
Rename module
